requirements.txt
opencv-python==4.8.1.78
google-generativeai==0.3.2
pillow==10.0.1
pygame==2.5.2
speechrecognition==3.10.0
pyttsx3==2.90
numpy==1.24.3
python-dotenv==1.0.0